+++ changelog.md
@@ -3,6 +3,7 @@
 ## Upcoming
 * Renamed equipment slot "feets" to "feet"
 * Weights can now be omitted and will default to 1
+* Range object can now contain "value" instead of "min"
 
 ## Beta 3.7.0
 * Renamed `on_hit_effect` to `events`